10-prompt-guardian/function/requirements.txt
azure-functions
openai>=1.0.0
azure-storage-blob>=12.0.0
azure-keyvault-secrets>=4.0.0
azure-identity>=1.0.0
